Move handleRandomIndex function to Home component
